@@ -8,9 +8,10 @@
     ],
     "homepage": "http://framework.zend.com/",
     "require": {
-        "php": ">=5.5",
+        "php": ">=7.0",
         "zendframework/zendframework": "~2.5",
-        "doctrine/orm": "^2.6",
-        "doctrine/cache": "^1.11"
+        "doctrine/orm": "^2.10",
+        "symfony/cache": "^5.4",
+        "doctrine/annotations": "^1.0"
     }
 }
